--- Assets/Tests/Common/Tests.Common.asmdef
@@ -1,3 +1,0 @@
-{
-	"name": "Tests.Common"
-}
