--- Assets/Tests/Common/Tests.Common.asmdef.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: d1b0850ddef09db439fe14a45bb2803c
-AssemblyDefinitionImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
