--- Assets/Tests/IntegrationTests.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 7394cb0ac99068a4b984c3ea0e4f1847
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
