--- Assets/Tests/UnitTests.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 90462ed347d20044d87af7b42a694b72
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
