+++ Assets/Tests/old.meta
@@ -1,6 +1,7 @@
 fileFormatVersion: 2
-guid: 256840ab2263ef3438ba1c02badda202
-AssemblyDefinitionImporter:
+guid: 8f06067451c62ed408c521646945969e
+folderAsset: yes
+DefaultImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
