+++ ProjectSettings/EditorBuildSettings.asset
@@ -15,7 +15,7 @@ EditorBuildSettings:
     path: Assets/Scenes/Game/Game.unity
     guid: 2bd5e08a742c0ab4d9db255b5c25d13f
   - enabled: 1
-    path: Assets/Tests/IntegrationTests/Scene/Tests scene.unity
+    path: Assets/Tests/PlayMode/Scene/Tests scene.unity
     guid: e58a48e65c191a2449beffb42f827ebf
   m_configObjects:
     com.unity.input.settings: {fileID: 11400000, guid: f6b25410aabc0e04ca2890a067e7b2b6, type: 2}
